Revert "update: organize components setup"
This reverts commit 8602d497c52f97accdbccf2c5393402fec2cdf34.
