+++ src/components/MainContent/Projects/ProjectFilter/ProjectFilter.jsx
@@ -1,8 +1,7 @@
-import { faSearch } from "@fortawesome/free-solid-svg-icons";
+import "./ProjectFilter.css";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-
-import "./ProjectFilter.css";
+import { faSearch } from "@fortawesome/free-solid-svg-icons";
 
 function Root({ children, onHandleSubmit }) {
   return (
