+++ src/components/MainContent/Projects/Projects.jsx
@@ -1,13 +1,12 @@
-import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
-import { intersectionService } from "../../../services/intersection-service";
+import "./Projects.css";
 
 import { ProjectModal } from "./ProjectModal/ProjectModal";
 import { ProjectCard } from "./ProjectCard/ProjectCard";
 import { ProjectFilter } from "./ProjectFilter/ProjectFilter";
 import { Modal } from "../../Modal/Modal";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-
-import "./Projects.css";
+import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
+import { intersectionService } from "../../../services/intersection-service";
 
 function Root({ children, onHandleKeyDown }) {
   return (
