+++ src/components/MainContent/Skills/Skills.jsx
@@ -1,15 +1,14 @@
-import { intersectionService } from "../../../services/intersection-service";
+import "./Skills.css";
+
 import {
   faArrowDown,
   faUpRightFromSquare,
 } from "@fortawesome/free-solid-svg-icons";
-
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Modal } from "../../Modal/Modal";
 import { SkillCard } from "./SkillCard/SkillCard";
 import { SkillModal } from "./SkillModal/SkillModal";
-
-import "./Skills.css";
+import { intersectionService } from "../../../services/intersection-service";
 
 function Root({ children, onHandleKeyDown }) {
   return (
