+++ src/main.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-
-import {App} from "./components/App/App";
-
+import App from "./components/App/App";
 import "./reset.css";
 import "./index.css";
 import "./interactions.css";
